refactor: useFetch 커스텀 훅 제작
